Fix hourly averages being set to the current average regardless of the unskewed average being present Add documentation Indentation improvements

+++ src/config/mod.rs
@@ -27,7 +27,7 @@ create_config_struct!(Spike,
 create_config_struct!(UnskewedAverage,
     reset_pcnt:      f32 => "Reset To Average Percentage"  => [0.0, 0.15],
     adjust_pcnt:     f32 => "Adjust to Average Percentage" => [0.0, 0.0075],
-    spikes_required: u8  => "Listener Spikes Required"     => 1,
+    spikes_required: u32  => "Listener Spikes Required"    => 1,
     jump_required:   f32 => "Listener Jump Required"       => [1.1, 4.0],
 );
